backend/requirements.txt
fastapi
uvicorn[standard]
pandas
sqlalchemy
psycopg2-binary
PyYAML
reportlab
